refactor: Updated imports to use hook index file, comment cleanup
